third_party/blink/web_tests/svg/hittest/ellipse-hittest-expected.txt

This tests hit testing on ellipses with continuous strokes. If this test passes, you will see "PASS" below.

PASS

(27, 46)
(98, 33)
(202, 53)
(98, 142)
(130, 180)
(91, 247)
(27, 240)
(336, 166)
(337, 214)
(31, 18)
(209, 31)
(132, 47)
(229, 43)
(245, 180)
(45, 180)
(130, 95)
(130, 295)
(212, 255)
(280, 235)
(301, 247)
(378, 88)
(335, 122)
(333, 190)
(377, 66)
